Fix pandas deprecation warnings in make_synth_data.py
- Replace .fillna(0.0) with .astype(float).fillna(0.0) to prevent FutureWarning
- Eliminates 2,728 pandas warnings from test suite
- Test results: 98 passed, 87 skipped, 0 failed (down from 2,730 warnings to 2 warnings)

+++ data_fetch/make_synth_data.py
@@ -214,8 +214,8 @@ def generate_prices(calendar: pd.DatetimeIndex, tickers: List[str],
                                 day_date=date_tm1)
         
         # Align proxies to active tickers; fill missing with 0
-        mom_z = mom_z.reindex(active).fillna(0.0)
-        val_z = val_z.reindex(active).fillna(0.0)
+        mom_z = mom_z.reindex(active).astype(float).fillna(0.0)
+        val_z = val_z.reindex(active).astype(float).fillna(0.0)
         
         # Generate components
         market_t = np.random.normal(loc=MARKET_MEAN, scale=MARKET_STD)
